move and refactor test to validate full integration within file logger and isx pipeline
