+++ version0/tests/isx_pipeline_test.py
@@ -1,6 +1,5 @@
 import unittest
 
-from isx_pipeline.isx_pipeline import ISXPipeline
 
 class ISXDouble:
     pass
